Move test folder to root directory
