+++ README.md
@@ -1,2 +1,62 @@
-# RPG-phaser
-My first game with the HTML5 framework Phaser
+# Karen's Adventure - Phaser RPG
+EasyWeather is a project in which I use OpenWeather API to display the temp and conditions of the weather of the desired city.
+
+This app is optimized for desktop, mobile isn't a good experience.
+
+![screenshot](dist/assets/img/screenshot.jpeg)
+
+## Live demo
+
+[live demo](https://rawcdn.githack.com/mricanho/weather-app/37a319bdc46aeddcc8164359331defec361fd294/dist/index.html)
+
+## Built With
+
+- JavaScript
+- HTML
+- CSS
+- Webpack
+- Package.json
+- OpenWeather API
+- Geocode API
+
+## How to start the project from your Local environment
+
+- Open your terminal and cd where you want to store the project
+- To clone the repository run the following command:
+```bash
+git clone https://github.com/mricanho/Restaurant-page.git
+```
+- Change directory to point to the Restaurant-page dir.
+- Now run:
+```bash
+npm install
+```
+- Finally run:
+```bash
+npm start
+```
+- Go to the dist folder and select index.html, now Open the live server on your VScode `Alt + L` then `Alt + o`
+
+## Author
+
+👤 **Miguel Ricaño**
+
+- Github: [@mricanho](https://github.com/mricanho)
+- Linkedin: [Miguel Ricaño](https://www.linkedin.com/in/mricanho/)
+
+## 🤝 Contributing
+
+Contributions, issues, and feature requests are welcome!
+
+## Show your support
+
+Give a star if you :star: like this project!
+
+## Acknowledgments
+
+- Microverse
+- GitHub
+- TheOdinProject
+- JavaScript
+
+This project is [MIT](LICENSE) licensed.
