+++ tests/Message.test.js
@@ -1,5 +1,5 @@
 import Phaser from 'phaser';
-import Message from '../modules/menus/message';
+import Message from '../src/modules/menus/message';
 
 test('Message is a subclass of container', () => {
   expect(Message.prototype instanceof Phaser.GameObjects.Container).toBe(true);
