+++ tests/actionsMenu.test.js
@@ -1,5 +1,5 @@
-import Menu from '../modules/menus/menu';
-import ActionsMenu from '../modules/menus/action';
+import Menu from '../src/modules/menus/menu';
+import ActionsMenu from '../src/modules/menus/action';
 
 test('ActionsMenu is a subclass of Menu', () => {
   expect(ActionsMenu.prototype instanceof Menu).toBe(true);
