+++ tests/enemy.test.js
@@ -1,5 +1,5 @@
 import Phaser from 'phaser';
-import Enemy from '../modules/characters/enemy';
+import Enemy from '../src/modules/characters/enemy';
 
 test('Enemy is a subclass of Sprite', () => {
   expect(Enemy.prototype instanceof Phaser.GameObjects.Sprite).toBe(true);
