+++ tests/player.test.js
@@ -1,5 +1,5 @@
 import Phaser from 'phaser';
-import Player from '../modules/characters/player';
+import Player from '../src/modules/characters/player';
 
 test('Player is a subclass of Sprite', () => {
   expect(Player.prototype instanceof Phaser.GameObjects.Sprite).toBe(true);
