+++ tests/unit.test.js
@@ -1,5 +1,5 @@
 import Phaser from 'phaser';
-import Unit from '../modules/characters/unit';
+import Unit from '../src/modules/characters/unit';
 
 test('Unit is a subclass of Sprite', () => {
   expect(Unit.prototype instanceof Phaser.GameObjects.Sprite).toBe(true);
